pedantic clippy

+++ casperfpga/src/yellow_blocks/mod.rs
@@ -6,5 +6,5 @@ pub mod ten_gbe;
 /// Certain Yellow Block struct types will implement this trait to allow for auto offsets in
 /// transport read methods
 pub trait Address {
-    fn addr() -> u8;
+    fn addr() -> u16;
 }
